+++ test/modules/Auction/MockBatchAuctionModule.sol
@@ -185,7 +185,7 @@ contract MockBatchAuctionModule is AuctionModule {
     function _claimProceeds(uint96 lotId_)
         internal
         override
-        returns (uint96, uint96, uint96, bool)
+        returns (uint256, uint96, uint96, bool)
     {
         // Update status
         lotStatus[lotId_] = Auction.Status.Settled;
